requirements.txt
fastapi
uvicorn
httpx
pydantic
python-dotenv
jinja2
python-multipart
